SampleImageScoring/modules/ScoreAndAnnotate/requirements.txt
azure-iot-device~=2.0.0
azure-core==1.7.0
azure-storage-blob==12.3.2
opencv-contrib-python
opencv-python
requests==2.20.1